+++ app/api/auth/[...nextauth]/route.ts
@@ -1,75 +1,5 @@
-import NextAuth, { AuthOptions } from "next-auth";
-import CredentialsProvider from "next-auth/providers/credentials";
-import prisma from "@/lib/prisma";
-import bcrypt from "bcryptjs";
-
-declare module "next-auth" {
-  interface Session {
-    user: {
-      id: string;
-      email: string;
-      name: string;
-    }
-  }
-}
-
-export const authOptions: AuthOptions = {
-  providers: [
-    CredentialsProvider({
-      name: "Credentials",
-      credentials: {
-        email: { label: "Email", type: "email" },
-        password: { label: "Password", type: "password" }
-      },
-      async authorize(credentials) {
-        if (!credentials?.email || !credentials?.password) {
-          return null;
-        }
-
-        const user = await prisma.user.findUnique({
-          where: {
-            email: credentials.email,
-          },
-        });
-
-        if (!user) {
-          return null;
-        }
-
-        const isPasswordValid = await bcrypt.compare(
-          credentials.password,
-          user.password
-        );
-
-        if (!isPasswordValid) {
-          return null;
-        }
-
-        return {
-          id: user.id.toString(),
-          email: user.email,
-          name: user.name,
-        };
-      }
-    })
-  ],
-  pages: {
-    signIn: '/login',
-  },
-  callbacks: {
-    session({ session, token }) {
-      if (session.user) {
-        session.user.id = token.sub!;
-        session.user.email = token.email!;
-        session.user.name = token.name!;
-      }
-      return session;
-    }
-  },
-  session: {
-    strategy: "jwt",
-  }
-};
+import NextAuth from "next-auth";
+import { authOptions } from "@/lib/auth.config";
 
 const handler = NextAuth(authOptions);
